microservices/cart-service/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
redis==5.0.8